requirements.txt
absl-py==2.1.0
aio-pika==9.5.4
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiormq==6.8.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==24.2.0
azure-core==1.33.0
azure-identity==1.21.0
azure-keyvault-secrets==4.9.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
cmdstanpy==1.2.5
colorama==0.4.6
comm==0.2.3
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
distlib==0.3.9
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.116.1
filelock==3.18.0
findspark==2.0.1
Flask==3.1.1
Flask-BasicAuth==0.2.0
Flask-HTTPAuth==4.8.0
Flask-SQLAlchemy==3.1.1
fonttools==4.54.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.3.0
future==1.0.0
gevent==24.11.1
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.73.0
h11==0.14.0
h5py==3.12.1
holidays==0.67
html5lib==1.1
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
iso8601==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.6.0
kiwisolver==1.4.7
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.0
MouseInfo==0.1.3
msal==1.32.0
msal-extensions==1.3.1
multidict==6.1.0
multitasking==0.0.11
namex==0.0.8
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.1.2
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
optree==0.13.0
outcome==1.3.0.post0
packaging==24.1
pamqp==3.3.0
pandas==2.2.3
paramiko==3.5.0
parso==0.8.5
peewee==3.17.7
pika==1.3.2
pillow==11.0.0
platformdirs==4.3.6
prometheus_client==0.22.1
prompt_toolkit==3.0.52
propcache==0.2.0
prophet==1.1.6
protobuf==5.29.5
psutil==7.0.0
psycopg2==2.9.10
pure_eval==0.2.3
puttykeys==1.0.3
py4j==0.10.9.9
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.10.1
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyodbc==5.2.0
pyparsing==3.2.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pyserial==3.5
pysftp==0.2.9
PySocks==1.7.1
pyspark==4.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==27.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
selenium==4.29.0
selenium-stealth==1.0.6
serial==0.0.97
setuptools==75.8.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.43
stack-data==0.6.3
stanio==0.5.1
starlette==0.47.1
textblob==0.19.0
threadpoolctl==3.5.0
tokenizers==0.21.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.0
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
undetected-chromedriver==3.5.5
urllib3==2.2.3
uvicorn==0.35.0
virtualenv==20.31.2
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket==0.2.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
yarl==1.18.0
yfinance==0.2.48
zabbix-api==0.5.6
zipp==3.23.0
zope.event==5.0
zope.interface==7.2
